+++ tests/ssp_loader_fmuproxy_test.cpp
@@ -12,23 +12,23 @@
 int main()
 {
     try {
-        cse::log::setup_simple_console_logging();
-        cse::log::set_global_output_level(cse::log::info);
+        cosim::log::setup_simple_console_logging();
+        cosim::log::set_global_output_level(cosim::log::info);
 
         const auto testDataDir = std::getenv("TEST_DATA_DIR");
         REQUIRE(testDataDir);
         boost::filesystem::path sspFile = boost::filesystem::path(testDataDir) / "ssp" / "demo" / "fmuproxy";
 
-        cse::ssp_loader loader;
+        cosim::ssp_loader loader;
         const auto config = loader.load(sspFile);
-        auto exec = cse::execution(config.start_time, config.algorithm);
-        const auto entityMaps = cse::inject_system_structure(
+        auto exec = cosim::execution(config.start_time, config.algorithm);
+        const auto entityMaps = cosim::inject_system_structure(
             exec,
             config.system_structure,
             config.parameter_sets.at(""));
         REQUIRE(entityMaps.simulators.size() == 2);
 
-        auto result = exec.simulate_until(cse::to_time_point(1e-3));
+        auto result = exec.simulate_until(cosim::to_time_point(1e-3));
         REQUIRE(result.get());
     } catch (const std::exception& e) {
         std::cerr << "Error: " << e.what();